+++ lib/zk/event_handler_subscription.rb
@@ -15,9 +15,21 @@ module ZK
     # @return [Proc]
     attr_accessor :callback
 
+    # an array of what kinds of events this handler is interested in receiving
+    #
+    # @return [Set] containing any combination of :create, :change, :delete,
+    #   or :children
+    #
     # @private
-    def initialize(event_handler, path, callback)
+    attr_accessor :interests
+
+    ALL_EVENTS    = [:created, :deleted, :changed, :child].freeze unless defined?(ALL_EVENTS)
+    ALL_EVENT_SET = Set.new(ALL_EVENTS).freeze                    unless defined?(ALL_EVENT_SET)
+
+    # @private
+    def initialize(event_handler, path, callback, interests)
       @event_handler, @path, @callback = event_handler, path, callback
+      @interests = prep_interests(interests)
     end
 
     # unsubscribe from the path or state you were watching
@@ -31,6 +43,26 @@ module ZK
     def call(event)
       callback.call(event)
     end
+
+    private
+      def prep_interests(a)
+        return ALL_EVENT_SET if a.nil?
+
+        rval = 
+          case a
+          when Array
+            Set.new(a)
+          when Symbol
+            Set.new([a])
+          else
+            raise ArgumentError, "Don't know how to handle interests: #{a.inspect}" 
+          end
+
+        rval.tap do |rv|
+          invalid = (rv - ALL_EVENT_SET)
+          raise ArgumentError, "Invalid event name(s) #{invalid.to_a.inspect} given" unless invalid.empty?
+        end
+      end
   end
 end